Prepare Person-Time and Event Data for SIR/IRR Calculations

This function merges exposure, response, and population-level data to produce an aggregated dataset for person-time and event rate analysis (e.g., Standardized Incidence Ratios or Incidence Rate Ratios). It allows comparison of disease occurrence across different exposure time windows.

Usage

pirr_data(
  d_exposure,
  d_response,
  d_population,
  var_age_start = 50,
  dg_list = NA,
  censoring_date = as.Date("2022-12-31"),
  limits = c(0.3, 3)
)

Arguments

d_exposure

A `data.frame` of exposure diagnoses with at least `ID`, `DATE`, and `DG` columns. Created by function `search_diagnoses()`.

d_response

A `data.frame` of response diagnoses with at least `ID`, `DATE`, and `DG` columns. Created by function `search_diagnoses()`. In case of parameter `dg_list` is given, give full list of diagnoses.

d_population

A `data.frame` containing baseline population data, including `ID`, `DATE_BIRTH`, `DATE_DEATH`, and `DATE_MIGRATION`. Created by function `classify_population()`.

var_age_start

Numeric. Age (in years) at which to start follow-up (e.g., 50).

dg_list

Optional named list. Each element is a string of diagnosis groupings (e.g., `"hip+forearm+humerus"`). If `NA`, the default "normal" procedure is applied.

censoring_date

A `Date` object indicating administrative censoring (e.g., `"2022-12-31"`).

limits

Numeric vector of length two. Optional plotting or analysis limits, default is `c(0.3, 3)`.

Value

A `data.frame` summarizing person-years (`pyrs`), event counts (`Death`), and optional diagnosis counts across exposure categories (`caika`) and age groups.

Details

The function:

Calculates first exposure and response dates per individual.

Computes time intervals between exposure and response.

Splits follow-up time using `heaven::lexisSeq()` and `heaven::lexisTwo()`.

Categorizes person-time into pre-defined exposure windows (e.g., `<1y`, `1-4y`, `5-9y`, etc.).

Optionally aggregates response diagnoses using `dg_list` expressions.

Filters follow-up to start at `var_age_start` (e.g., 50 years).

Examples

if (FALSE) { # \dontrun{
result <- pirr_data(
  d_exposure = exposure_data,
  d_response = response_data,
  d_population = population_data,
  var_age_start = 50,
  censoring_date = as.Date("2023-12-31")
)
} # }
if (FALSE) { # \dontrun{
result <- pirr_data(
  d_exposure = exposure_data,
  d_response = diagnoses %>% filter(DGREG == "FRACTURES" ),
  d_population = dpop,
  dg_list = list(any_fracture = "ankle+forearm+hip+humerus+vertebral",
              ostheoporotic = "forearm+hip+humerus+vertebral"),
  var_age_start = 50,
  censoring_date = as.Date("2023-12-31")
)
} # }